lib/serf: fix a Serf init bug as well as minor issues
When initializing Serf, do not set memberlist.AdvertiseAddr, as it's
not necessary. It only causes other bugs that a peer address/port
becomes false when a new peer joins.

Store each Serf DB under a directory with a unique name, instead of
common directory, to avoid conflicts when creating DBs.

Fix minor fault in a debugging message.