raft: implement FSM with simple events slice
We can use the raft FSM to replicate state from the Leader to other
nodes of the cluster. In this way, when a new Follower becomes Leader it
has all the state needed.

This is an example implementation, it will become the event queue to
store unsent push events to the Kyma cluster and will allow message
delivery guarantee with retries in case of transport failure.