connector: add proxy capabilities
We implement an HTTPS/2.0 and HTTPS/1.1 proxy by handling CONNECT
requests and proxying the traffic to the Kyma server.

If there's no existing connection to Kyma, it will be
established and if there is, it will be reused.

For HTTPS/1 connections, we hijack them and send them through the
HTTPS/2.0 tunnel established between wormhole-connector and the Kyma
server.

For HTTPS/2 connections, we don't need to (and can't) hijack them, we
can just stream directly.

This also defines the kyma-server parameter like it should be: a remote
address to redirect proxied traffic to.

We should still define timeouts, make sure connections are closed
cleanly and clean things up, but the basic functionality is there.

Also, this only works if requests are sent to the wormhole-connector
cluster leader.