internal/*,examples: Implement bidirectional connectivity
With this the Wormhole Dispatcher can also act as a proxy for
applications inside the Kyma cluster that will send traffic to the
Wormhole Connector.

As a first stab, this is implemented as two persistent HTTP/2
connections, one for each direction. Also, the Wormhole Dispatcher needs
to know the address of the Wormhole Connector.

Future improvements:

* Establish the two tunnels from the Wormhole Connector (so reversing
client/server for the connection Dispatcher -> Connector) so the
Dispatcher doesn't need to know the address of the Connector.
* Use only an HTTP/2 connection.